pyproject.toml
[project]
name = "adk-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.5.0",
    "google-cloud-bigquery>=3.34.0",
    "google-cloud-documentai>=3.5.0",
    "google-cloud-storage>=2.19.0",
    "langchain>=0.3.26",
    "langchain-google-genai>=2.1.7",
    "numpy>=2.3.1",
    "pandas>=2.3.0",
    "python-dotenv>=1.1.1",
]
